move ImgUtils to util package
